refactor: 일별 회원가입 통계 수정
