+++ src/main/java/com/phcworld/repository/statistics/StatisticsRepository.java
@@ -2,6 +2,7 @@ package com.phcworld.repository.statistics;
 
 import com.phcworld.domain.board.QDiary;
 import com.phcworld.domain.statistics.StatisticsDto;
+import com.phcworld.domain.user.QUser;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.core.types.dsl.StringTemplate;
@@ -18,6 +19,43 @@ import java.util.List;
 public class StatisticsRepository {
     private final JPAQueryFactory queryFactory;
     QDiary diary = QDiary.diary;
+
+    QUser user = QUser.user;
+
+    public List<StatisticsDto> findRegisterUserStatistics(LocalDate startDate, LocalDate endDate){
+        LocalDateTime start = startDate.atStartOfDay();
+        LocalDateTime end = endDate.atTime(23,59,59);
+        // mysql
+//        DateTemplate<String> formattedDate = Expressions.dateTemplate(
+//                String.class,
+//                "DATE_FORMAT({0}, {1})",
+//                user.createDate,
+//                ConstantImpl.create("%Y-%m-%d")
+//        );
+        StringTemplate formattedDate = Expressions.stringTemplate(
+                "DATE_FORMAT({0}, {1})",
+                user.createDate,
+//                ConstantImpl.create("%Y-%m-%d")
+                "%Y-%m-%d"
+        );
+
+//        StringTemplate formattedDate = Expressions.stringTemplate(
+//                "FORMATDATETIME({0}, {1})",
+//                user.createDate,
+//                "yyyy-MM-dd"
+//        );
+
+
+        return queryFactory.select(Projections.fields(StatisticsDto.class,
+                        user.count().as("count"),
+                        formattedDate.as("date")))
+                .from(user)
+                .where(user.createDate.between(start, end))
+//                        user.createDate.goe(start)
+//                        .and(user.createDate.loe(end)))
+                .groupBy(formattedDate)
+                .fetch();
+    }
 
     public List<StatisticsDto> findPostDiaryStatistics(LocalDate startDate, LocalDate endDate){
         LocalDateTime start = startDate.atStartOfDay();
