+++ src/main/java/com/phcworld/repository/user/UserRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<User, Long>, UserRepositoryCustom {
+public interface UserRepository extends JpaRepository<User, Long> {
 	User findByEmail(String email);
 
 	@Query(nativeQuery = true,
