--- src/main/java/com/phcworld/repository/user/UserRepositoryCustomImpl.java
@@ -1,57 +1,0 @@
-package com.phcworld.repository.user;
-
-import com.phcworld.domain.statistics.StatisticsDto;
-import com.phcworld.domain.user.QUser;
-import com.querydsl.core.types.Projections;
-import com.querydsl.core.types.dsl.Expressions;
-import com.querydsl.core.types.dsl.StringTemplate;
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Repository
-@RequiredArgsConstructor
-public class UserRepositoryCustomImpl implements UserRepositoryCustom{
-    private final JPAQueryFactory queryFactory;
-    QUser user = QUser.user;
-
-    @Override
-    public List<StatisticsDto> findRegisterUserStatistics(LocalDate startDate, LocalDate endDate){
-        LocalDateTime start = startDate.atStartOfDay();
-        LocalDateTime end = endDate.atTime(23,59,59);
-        // mysql
-//        DateTemplate<String> formattedDate = Expressions.dateTemplate(
-//                String.class,
-//                "DATE_FORMAT({0}, {1})",
-//                user.createDate,
-//                ConstantImpl.create("%Y-%m-%d")
-//        );
-        StringTemplate formattedDate = Expressions.stringTemplate(
-                "DATE_FORMAT({0}, {1})",
-                user.createDate,
-//                ConstantImpl.create("%Y-%m-%d")
-                "%Y-%m-%d"
-        );
-
-//        StringTemplate formattedDate = Expressions.stringTemplate(
-//                "FORMATDATETIME({0}, {1})",
-//                user.createDate,
-//                "yyyy-MM-dd"
-//        );
-
-
-        return queryFactory.select(Projections.fields(StatisticsDto.class,
-                        user.count().as("count"),
-                        formattedDate.as("date")))
-                .from(user)
-                .where(user.createDate.between(start, end))
-//                        user.createDate.goe(start)
-//                        .and(user.createDate.loe(end)))
-                .groupBy(formattedDate)
-                .fetch();
-    }
-}
