+++ src/test/java/com/phcworld/service/statistics/StatisticsServiceTest.java
@@ -24,9 +24,10 @@ public class StatisticsServiceTest {
         LocalDate startDate = LocalDate.of(2023, 1, 1);
         LocalDate endDate = LocalDate.of(2023, 12, 31);
 
-//        List<StatisticsDto> list = statisticsService.getRegisterMemberStatistics(startDate, endDate);
+        List<StatisticsDto> userRegsiterList = statisticsService.getRegisterMemberStatistics(startDate, endDate);
+        log.info("user list : {}", userRegsiterList);
 //        List<UserStatistics> list = statisticsService.getRegisterMemberStatisticsForNativeQuery(startDate, endDate);
-        List<StatisticsDto> list = statisticsService.getPostDiaryStatistics(startDate, endDate);
-        log.info("list : {}", list);
+        List<StatisticsDto> diaryPostList = statisticsService.getPostDiaryStatistics(startDate, endDate);
+        log.info("diary list : {}", diaryPostList);
     }
 }
